changed default widget layout to consider non-widescreens
The left sidebar (tool dockwidgets) increases the program's minimum
height; and it doesn't leave much drawing space on a non-widescreen. I
might make an action for toggling all of them as a group.

The pen and option toolbars work well together; the option toolbar has
plenty of room to display anything it needs to, and the pen widget is
slightly shorter than regular toolbars, which makes the "line" that they
are on shorter than the others.