added widget options dialog & menu text size option
This demonstrates a different approach to an options dialog.
The dialog itself does nothing;
everything is located in QC_ApplicationWindow::widgetOptionsDialog
which eliminates having to read the settings directly after writing
them.

I acknowledge that rs_settings takes care of qvariant conversion and
uses a cache map;
however, I want to write this as modular as possible for now.

The values passed to QCoreApplication::setOrganizationName and
QCoreApplication::setApplicationName are remembered by qsettings.

Segregating the widget options makes it more modular for other projects;
and it causes the window to load faster.

I will add the statusbar height and text size option soon.