simplified splash screen code
- I could not find a valid reason for using #ifdef QSPLASHSCREEN_H
- The splash variable is now recognized by Qt Creator.
- Reimplementing QMainWindow::show() for using QTimer::singleShot is
unnecessary, and makes the splash screen dependent on it.

The splash screen is designed to indicate that the program is loading
prior to the display of the main window; and it is designed to be closed
via splash->finish(main_window).

This is prep-work for a "Start with main window maximized" option
(coming soon).