improved & moved statusbar font size option; issue #638
Removing the set sizes from the UI files allows for children of the
status bar to be affected by status_bar->setFont()
It also allows for the system default font to be used and properly scale
with the statusbar on first load.

I also added some space to the left of the coordinates to make it less
cramped.