removed devicePixelRatio from pen toolbar
This was producing a lot of extra space to the right of the toolbar
that could not be removed by resizing (on OS X 10.11.3 with a retina
display).
The toolbar is perfectly sized after removing this code.

I recommend allowing widgets to be automatically sized for optimal
cross platform results. If it can’t be achieved that way, then the
widget likely needs to be redesigned.