removed mac style sheet from main.cpp
I can see why this was added;
however, now that we have the option to choose different styles and
style sheets it makes sense to let the default style be shown.
Otherwise, when someone adds their own style sheet they might be
confused as to why this behavior disappears.
We can add this style sheet to the new resources repository.