Fix: limit scrolling to drawing + perimeter.
Before this patch, the scroll limits were arbitrary, which
made the user have to use zoom to recenter the scroll perimeter,
in order to navigate the whole document.

This patch limits scrolling to the drawing instead. The exact perimeters
are now so that the drawing is always in the viewport, with an extra
empty zone of 75% for potentially extending the drawing.

This issue got mentioned in #479