Caches the value of QScrollBar::sizeHint().
This method got called each time the window was drawn. This leads to a
chain of calls into the underlying windowing system (even into GTK, on
my machine). It took ~57% of an RS_Line::draw call.

Now the value gets cached, and the resizeEvent in QScrollBar is used to
update the cached value.