Changed user interface
It's more common to use addendum & dedendum parameters relative
to the modulus (which is the amount the pitch_diameter grows
as you add one tooth to your gear)

As all standard conformed gears of same modulus are compatible,
it's better to select the modulus and allow the program to calculate
the right pitch diameter, as normally yo design gears to interconnect
them, so better to select a modulus, then change the number of teeths
only.

So, I changed the pitch radius for the modulus, instead.

I have taken the liberty to also add my name to the credits line.

Signed-off-by: Luis Colorado <luiscoloradourcola@gmail.com>