fix-new-layer-auto-numbering
This allows for any base layer name with a suffix number to be auto
incremented. If no name is supplied it reverts to the default base layer
name of 'noname' and with any additional layers being provided a suffix
number automatically. It is also possible to just use numbered layers
with no base layer name and it will still auto increment correctly...the
only exception being layer '0'. So if you want simple numbered layers
click the + button then replace the default 'noname' layer name in the
dialog with '1' and click 'OK'. From then on highlighting the layer you
want to increment and clicking the '+' button will give you a new layer
with the next available number above the one you just selected.